LayoutTests/platform/chromium/http/tests/inspector/resource-tree/appcache-manifest-with-non-existing-file-expected.txt

Tests that application cache model keeps track of manifest urls and statuses correctly when there is a non existing file listed in manifest.

Bug 64581
Dumping application cache tree:
    (empty)
Dumping application cache model:
    (empty)

Dumping application cache tree:
    (empty)
Dumping application cache model:
    (empty)

Dumping application cache statuses records:
    Frame: frame1
        manifest url: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=with-non-existing-file
        status:       CHECKING
        manifest url: http://127.0.0.1:8000/inspector/resource-tree/resources/manifest.php?manifestId=with-non-existing-file
        status:       DOWNLOADING
        manifest url: 
        status:       UNCACHED
